Switched over every sqrt() to a sqrtf() for speed reasons.

+++ TDML/TDMLmath.cpp
@@ -52,7 +52,7 @@ namespace TDML
 
 	float TDMLmath::distance(float x1, float y1, float z1, float x2, float y2, float z2)
 	{
-		return sqrt(((x1-x2)*(x1-x2))+((y1-y2)*(y1-y2))+((z1-z2)*(z1-z2)));
+		return sqrtf(((x1-x2)*(x1-x2))+((y1-y2)*(y1-y2))+((z1-z2)*(z1-z2)));
 	}
 
 	float TDMLmath::radToDegree(float radians)
